% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{corner}
\alias{corner}
\title{Prints the top left corner of a matrix}
\usage{
corner(X, n = 10)
}
\arguments{
\item{X}{matrix or data.frame}

\item{n}{num of columns and rows to print}
}
\value{
Prints the topleft corner of a matrix
}
\description{
This function prints the topleft corner of a matrix.
}
\examples{
corner(X = matrix)
}
