% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect.iterate.R
\name{intersect.iterate}
\alias{intersect.iterate}
\title{Performs iterative intersect comparisons to find the intersect of >2 vectors.}
\usage{
intersect.iterate(list_of_vectors)
}
\arguments{
\item{list_of_vectors}{list object containing the target vectors}
}
\value{
vector of the items in common between all vectors in list_of_vectors
}
\description{
This function will find the items in common between a set of >2 vectors.
}
\examples{
X <- list( LETTERS[1:15], LETTERS[3:18], LETTERS[6:23] )
intersect.iterate(X)
}
