% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na.omit.vector.R
\name{na.omit.vector}
\alias{na.omit.vector}
\title{Removes NAs from a vector}
\usage{
\method{na.omit}{vector}(x)
}
\arguments{
\item{x}{vector}
}
\value{
Prints the topleft corner of a matrix
}
\description{
This function removes NA values from a vector
}
\examples{
na.omit.vector(x = vector)
}
