% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Removes outliers from a vector of numbers}
\usage{
remove_outliers(x)
}
\arguments{
\item{x}{vector of numbers}
}
\description{
This function removes outliers using IQR +/- the 75th and 25th percentile. This is the same calculation as the boxplot() function.
}
\examples{
remove_outliers(x = vector)
}
