% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_center_matrix.R
\name{scale_center_matrix}
\alias{scale_center_matrix}
\title{Scale and center a matrix}
\usage{
scale_center_matrix(X, MARGIN = 1)
}
\arguments{
\item{X}{matrix or data.frame}

\item{MARGIN}{scale and center row-wise (1; default) or column-wise (2)}
}
\value{
Returns the scaled and centered matrix
}
\description{
This function will linearly scale the row or column values in a matrix between 0 and 1, and then subtract the scaled mean.
}
\examples{
X <- matrix(data = 1:80, nrow = 10)
scale_center_matrix(X = X, MARGIN = 1)
}
