% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_by_group.R
\name{standardize_by_group}
\alias{standardize_by_group}
\title{Standardize a numeric vector based on group identity}
\usage{
standardize_by_group(x, groups)
}
\arguments{
\item{x}{numeric vector}

\item{groups}{vector with group labels}
}
\value{
Returns a numeric vector standardized by group label.
}
\description{
This function scales and centers a numeric vector within groups. Numbers from a single group are scaled linearly between 0 and 1, then centered by subtracting the scaled mean. This process is performed idenpendently for all groups.
}
\examples{
x <- 9:1
groups <- c("A", "B", "C", "A", "B", "C", "A", "B", "C" )
standardize_by_group(x = x, groups = groups)
> 0.5  0.5  0.5  0.0  0.0  0.0 -0.5 -0.5 -0.5
}
