% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transparent_color.R
\name{transparent_color}
\alias{transparent_color}
\title{Converts a named color to a transparent version.}
\usage{
transparent_color(color_name, alpha)
}
\arguments{
\item{color_name}{color name recognized by R}

\item{alpha}{alpha level: value between 0 and 1, smaller values indicate more transparency}
}
\value{
RGD code of the color with transparency included
}
\description{
This function takes a named color and reduced the opacity based on the alpha argument.
}
\examples{
transparent_color(color_name = "red", alpha = 0.25)
}
