% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation_by_group.R
\name{variation_by_group}
\alias{variation_by_group}
\title{Calculate the variantion in a numeric vector explained by group identity}
\usage{
variation_by_group(x, group_labels)
}
\arguments{
\item{x}{numeric vector}

\item{group_labels}{vector with group labels}
}
\value{
Returns the proportion of variation explained by provided group labels.
}
\description{
This function calculates the variation in a vector that is explained by group labels using lm() and anova().
}
\examples{
x <- 9:1
group_labels <- c("A", "B", "C", "A", "B", "C", "A", "B", "C" )
variation_by_group(x = x, group_labels = group_labels)
}
