% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df.R
\name{write_df}
\alias{write_df}
\title{Saves matrix/dataframe to tab-delimited text file.}
\usage{
write_df(X, rowNm_header, file)
}
\arguments{
\item{X}{matrix or data.frame}

\item{rowNm_header}{Column ID for row names}

\item{file}{Location to save file}
}
\value{
Writes the matrix/data.frame to text file
}
\description{
This function saves matrix/dataframe to tab-delimited text file with an ID included for the row names column
}
\examples{
write_df(X = matrix, rowNm_header = "rowType", file = "../path/to/file.txt")
}
