% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{corner}
\alias{corner}
\title{Print the corner of a matrix}
\usage{
corner(X, n = 6, c = "topleft")
}
\arguments{
\item{X}{matrix or data.frame}

\item{n}{num of columns and rows to print}

\item{c}{matrix corner to print: "topleft" (default), "topright", "bottomleft", and "bottomright"}
}
\value{
Prints the corner of a matrix
}
\description{
This function prints a corner of a matrix. Corner to be printed is assigned with the 'c' parameter (default = top left).
}
\examples{
X <- matrix(data = 1:80, nrow = 10)

corner(X = X, n = 2, c = "topleft")
>      [,1] [,2]
> [1,]    1   11
> [2,]    2   12

corner(X = X, n = 3, c = "bottomright")
>      [,1] [,2] [,3]
> [1,]   58   68   78
> [2,]   59   69   79
> [3,]   60   70   80
}
